fix req 1 to 8

+++ src/controller/recipes/removeController.js
@@ -3,9 +3,10 @@ const service = require('../../service/recipes');
 module.exports = async (req, res, next) => {
   try {
     const { id } = req.params;
-    const result = await service.removeService(id);
+    const { _id, role } = req.user;
+    const result = await service.removeService({ id, userId: _id, role });
 
-    return res.status(200).json(result);
+    return res.status(204).json(result);
   } catch (error) {
     next(error);
   }
